requirements.txt
streamlit
google-cloud-storage
google-generativeai
ffmpeg-python
tenacity
nest_asyncio
fastapi
uvicorn
python-dotenv
pytest-playwright
playwright